refactor: split process output polling
